refactor(catalog): decentralize entity filter mappings per plugin
Move entity type property maps and filter logic from the centralized
catalog/internal/db/filter/entity_mappings.go into per-plugin service
packages. Each GenericRepository only queries its own entity type, so
the centralized routing was unnecessary indirection that forced shared
file edits when adding new plugins.
