--- catalog/internal/db/filter/entity_mappings.go
@@ -1,223 +1,0 @@
-package filter
-
-import (
-	"strings"
-
-	catalogmodels "github.com/kubeflow/hub/catalog/internal/db/models"
-	"github.com/kubeflow/hub/internal/platform/db/filter"
-)
-
-// catalogEntityMappings implements EntityMappingFunctions for the catalog package
-type catalogEntityMappings struct{}
-
-// NewCatalogEntityMappings creates a new instance of catalog entity mappings
-func NewCatalogEntityMappings() filter.EntityMappingFunctions {
-	return &catalogEntityMappings{}
-}
-
-// GetMLMDEntityType maps catalog REST entity types to their underlying MLMD entity type
-func (c *catalogEntityMappings) GetMLMDEntityType(restEntityType filter.RestEntityType) filter.EntityType {
-	switch restEntityType {
-	case filter.RestEntityType(catalogmodels.RestEntityCatalogArtifact):
-		return filter.EntityTypeArtifact
-	case filter.RestEntityType(catalogmodels.RestEntityMCPServerTool):
-		return filter.EntityTypeExecution
-	default:
-		return filter.EntityTypeContext
-	}
-}
-
-// GetPropertyDefinitionForRestEntity returns property definition for a catalog REST entity type
-func (c *catalogEntityMappings) GetPropertyDefinitionForRestEntity(restEntityType filter.RestEntityType, propertyName string) filter.PropertyDefinition {
-	// Check if this is a well-known property for catalog entities
-	if restEntityType == filter.RestEntityType(catalogmodels.RestEntityCatalogModel) {
-		if _, isWellKnown := catalogModelProperties[propertyName]; isWellKnown {
-			// Use the well-known property definition
-			return catalogModelProperties[propertyName]
-		}
-
-		// Check if this is a property path referencing a related artifact
-		// Format: artifacts.<propertyName> or artifacts.customProperties.<propertyName>
-		if artifactPropertyPath, found := strings.CutPrefix(propertyName, "artifacts."); found {
-			// Return a RelatedEntity property definition
-			// ValueType is left empty to allow runtime type inference from the value
-			return filter.PropertyDefinition{
-				Location:          filter.RelatedEntity,
-				ValueType:         "", // Empty to enable runtime type inference
-				Column:            artifactPropertyPath,
-				RelatedEntityType: filter.RelatedEntityArtifact,
-				RelatedProperty:   artifactPropertyPath,
-				JoinTable:         "Attribution", // Join through Attribution table
-			}
-		}
-	}
-
-	if restEntityType == filter.RestEntityType(catalogmodels.RestEntityCatalogArtifact) {
-		if _, isWellKnown := catalogArtifactProperties[propertyName]; isWellKnown {
-			// Use the well-known property definition
-			return catalogArtifactProperties[propertyName]
-		}
-	}
-
-	if restEntityType == filter.RestEntityType(catalogmodels.RestEntityMCPServer) {
-		if _, isWellKnown := mcpServerProperties[propertyName]; isWellKnown {
-			// Use the well-known property definition
-			return mcpServerProperties[propertyName]
-		}
-	}
-
-	if restEntityType == filter.RestEntityType(catalogmodels.RestEntityMCPServerTool) {
-		if _, isWellKnown := mcpServerToolProperties[propertyName]; isWellKnown {
-			// Use the well-known property definition
-			return mcpServerToolProperties[propertyName]
-		}
-	}
-
-	// Not a well-known property for this entity type, treat as custom
-	return filter.PropertyDefinition{
-		Location:  filter.Custom,
-		ValueType: filter.StringValueType, // Default, will be inferred at runtime
-		Column:    propertyName,           // Use the property name as-is for custom properties
-	}
-}
-
-// IsChildEntity returns true if the catalog REST entity type uses prefixed names (parentId:name)
-func (c *catalogEntityMappings) IsChildEntity(entityType filter.RestEntityType) bool {
-	return false
-}
-
-// GetEqualityExpansion implements filter.EqualityExpander so that:
-//   - externalId = "x" matches both exact and namespaced (sourceId:x) stored values.
-//   - name = "modelName" matches both exact and namespaced (sourceId:modelName) stored values,
-//     so callers can filter by model name without requiring the source prefix.
-func (c *catalogEntityMappings) GetEqualityExpansion(restEntityType filter.RestEntityType, propertyName string, value any) (likeArg any, useExpansion bool) {
-	if restEntityType != filter.RestEntityType(catalogmodels.RestEntityCatalogModel) {
-		return nil, false
-	}
-	strVal, ok := value.(string)
-	if !ok || strVal == "" {
-		return nil, false
-	}
-	switch propertyName {
-	case "externalId":
-		// Match any source prefix: sourceId:externalId
-		return "%:" + escapeLike(strVal), true
-	case "name":
-		// If value already contains ":", treat as full namespaced name — exact match only.
-		if strings.Contains(strVal, ":") {
-			return nil, false
-		}
-		// Match namespaced form so model name without source prefix still finds the row.
-		return "%:" + escapeLike(strVal), true
-	default:
-		return nil, false
-	}
-}
-
-// escapeLike escapes SQL LIKE metacharacters (%, _, \) for safe use as a literal in a LIKE pattern.
-func escapeLike(s string) string {
-	var b strings.Builder
-	for _, r := range s {
-		switch r {
-		case '\\', '%', '_':
-			b.WriteRune('\\')
-			b.WriteRune(r)
-		default:
-			b.WriteRune(r)
-		}
-	}
-	return b.String()
-}
-
-// catalogModelProperties defines the allowed properties for CatalogModel entities
-var catalogModelProperties = map[string]filter.PropertyDefinition{
-	// Common Context properties
-	"id":                       {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "id"},
-	"name":                     {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "name"},
-	"externalId":               {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "external_id"},
-	"createTimeSinceEpoch":     {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "create_time_since_epoch"},
-	"lastUpdateTimeSinceEpoch": {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "last_update_time_since_epoch"},
-
-	// CatalogModel-specific properties stored in ContextProperty table
-	"source_id":      {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "source_id"},
-	"description":    {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "description"},
-	"owner":          {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "owner"},
-	"state":          {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "state"},
-	"language":       {Location: filter.PropertyTable, ValueType: filter.ArrayValueType, Column: "language"},
-	"library_name":   {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "library_name"},
-	"license_link":   {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "license_link"},
-	"license":        {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "license"},
-	"logo":           {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "logo"},
-	"maturity":       {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "maturity"},
-	"provider":       {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "provider"},
-	"readme":         {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "readme"},
-	"tasks":          {Location: filter.PropertyTable, ValueType: filter.ArrayValueType, Column: "tasks"},
-	"tags":           {Location: filter.PropertyTable, ValueType: filter.ArrayValueType, Column: "tags"},
-	"verifiedSource": {Location: filter.PropertyTable, ValueType: filter.BoolValueType, Column: "verifiedSource"},
-	"validated_tasks":  {Location: filter.PropertyTable, ValueType: filter.ArrayValueType, Column: "validated_tasks"},
-	"serving_config":   {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "serving_config"},
-}
-
-// catalogArtifactProperties defines the allowed properties for CatalogArtifact entities
-var catalogArtifactProperties = map[string]filter.PropertyDefinition{
-	// Common Artifact properties
-	"id":                       {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "id"},
-	"name":                     {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "name"},
-	"externalId":               {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "external_id"},
-	"createTimeSinceEpoch":     {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "create_time_since_epoch"},
-	"lastUpdateTimeSinceEpoch": {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "last_update_time_since_epoch"},
-	"uri":                      {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "uri"},
-	"state":                    {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "state"},
-
-	// Artifact type (stored in type_id but we can filter by string representation)
-	"artifactType": {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "artifactType"},
-}
-
-// mcpServerProperties defines the allowed properties for MCPServer entities.
-var mcpServerProperties = map[string]filter.PropertyDefinition{
-	// Common Context properties (Entity Table - required)
-	"id":                       {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "id"},
-	"name":                     {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "base_name"},
-	"externalId":               {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "external_id"},
-	"createTimeSinceEpoch":     {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "create_time_since_epoch"},
-	"lastUpdateTimeSinceEpoch": {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "last_update_time_since_epoch"},
-
-	// MCPServer-specific properties stored in ContextProperty table
-	"source_id":        {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "source_id"},
-	"base_name":        {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "base_name"},
-	"description":      {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "description"},
-	"provider":         {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "provider"},
-	"license":          {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "license"},
-	"license_link":     {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "license_link"},
-	"logo":             {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "logo"},
-	"readme":           {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "readme"},
-	"version":          {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "version"},
-	"tags":             {Location: filter.PropertyTable, ValueType: filter.ArrayValueType, Column: "tags"},
-	"transports":       {Location: filter.PropertyTable, ValueType: filter.ArrayValueType, Column: "transports"},
-	"deploymentMode":   {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "deploymentMode"},
-	"documentationUrl": {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "documentationUrl"},
-	"repositoryUrl":    {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "repositoryUrl"},
-	"sourceCode":       {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "sourceCode"},
-	"publishedDate":    {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "publishedDate"},
-	"lastUpdated":      {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "lastUpdated"},
-	"verifiedSource":   {Location: filter.PropertyTable, ValueType: filter.BoolValueType, Column: "verifiedSource"},
-	"secureEndpoint":   {Location: filter.PropertyTable, ValueType: filter.BoolValueType, Column: "secureEndpoint"},
-	"sast":             {Location: filter.PropertyTable, ValueType: filter.BoolValueType, Column: "sast"},
-	"readOnlyTools":    {Location: filter.PropertyTable, ValueType: filter.BoolValueType, Column: "readOnlyTools"},
-	"endpoints":        {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "endpoints"},
-	"artifacts":        {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "artifacts"},
-	"runtimeMetadata":  {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "runtimeMetadata"},
-}
-
-// mcpServerToolProperties defines the allowed properties for MCPServerTool entities
-var mcpServerToolProperties = map[string]filter.PropertyDefinition{
-	// Common Execution properties (Entity Table - required)
-	"id":                       {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "id"},
-	"name":                     {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "name"},
-	"createTimeSinceEpoch":     {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "create_time_since_epoch"},
-	"lastUpdateTimeSinceEpoch": {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "last_update_time_since_epoch"},
-
-	// MCPServerTool-specific properties stored in ExecutionProperty table
-	"description": {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "description"},
-	"accessType":  {Location: filter.PropertyTable, ValueType: filter.StringValueType, Column: "accessType"},
-}
