+++ tools/catalog-gen/generator_test.go
@@ -62,6 +62,8 @@ func TestGeneratorCreateFiles(t *testing.T) {
 		"catalog/internal/catalog/testcatalog/db_test.go",
 		"catalog/internal/catalog/testcatalog/models/test_entity.go",
 		"catalog/internal/catalog/testcatalog/service/test_entity.go",
+		"catalog/internal/catalog/testcatalog/service/test_entity_entity_mappings.go",
+		"catalog/internal/catalog/testcatalog/service/test_entity_entity_mappings_test.go",
 		"api/openapi/src/plugins/test.yaml",
 		"catalog/plugins/test/.openapi-generator-ignore",
 		"catalog/plugins/test/scripts/gen_openapi_server.sh",
