+++ tools/catalog-gen/templates/entity_mappings.go.tmpl
@@ -1,0 +1,48 @@
+package service
+
+import "github.com/kubeflow/hub/internal/platform/db/filter"
+
+type {{.Entity.CamelName}}EntityMappings struct{}
+
+// Unexported: only used by the repository constructor in this package.
+// Export (New...) if tests in a parent package need to call it directly.
+func new{{.Entity.Name}}EntityMappings() filter.EntityMappingFunctions {
+	return &{{.Entity.CamelName}}EntityMappings{}
+}
+
+func (m *{{.Entity.CamelName}}EntityMappings) GetMLMDEntityType(_ filter.RestEntityType) filter.EntityType {
+{{- if eq .Entity.DatastoreType "artifact"}}
+	return filter.EntityTypeArtifact
+{{- else if eq .Entity.DatastoreType "execution"}}
+	return filter.EntityTypeExecution
+{{- else}}
+	return filter.EntityTypeContext
+{{- end}}
+}
+
+func (m *{{.Entity.CamelName}}EntityMappings) GetPropertyDefinitionForRestEntity(_ filter.RestEntityType, propertyName string) filter.PropertyDefinition {
+	if def, ok := {{.Entity.CamelName}}Properties[propertyName]; ok {
+		return def
+	}
+	return filter.PropertyDefinition{
+		Location:  filter.Custom,
+		ValueType: filter.StringValueType,
+		Column:    propertyName,
+	}
+}
+
+func (m *{{.Entity.CamelName}}EntityMappings) IsChildEntity(_ filter.RestEntityType) bool {
+	return false
+}
+
+var {{.Entity.CamelName}}Properties = map[string]filter.PropertyDefinition{
+	"id":                       {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "id"},
+	"name":                     {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "name"},
+	"externalId":               {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "external_id"},
+	"createTimeSinceEpoch":     {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "create_time_since_epoch"},
+	"lastUpdateTimeSinceEpoch": {Location: filter.EntityTable, ValueType: filter.IntValueType, Column: "last_update_time_since_epoch"},
+{{- if eq .Entity.DatastoreType "artifact"}}
+	"uri":                      {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "uri"},
+	"state":                    {Location: filter.EntityTable, ValueType: filter.StringValueType, Column: "state"},
+{{- end}}
+}
